course_project/version_two/requirements.txt
langchain-huggingface==0.1.0
langchain-community==0.3.1
langchain-groq==0.2.0
langchainhub==0.1.21
faiss-cpu==1.8.0
python-dotenv
ragatouille
langgraph
pymupdf
